# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Native Observability & Threat Detection Engine for Global Payment APIs",
    "description": "Enterprise-grade runtime security, prompt injection detection, financial logic validation, and PCI DLP redaction engine for autonomous payment agents.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/transactions": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Process Payment Transaction",
        "description": "Submits a payment instruction through the full AI-Native layered security engine.\nPipeline: Prompt Scan -> Financial Policy Check -> Gate Validation -> DLP -> Risk Engine -> Ledger.",
        "operationId": "create_transaction_api_v1_transactions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "List Recent Transactions",
        "description": "Retrieves paginated transactions from the audit ledger.",
        "operationId": "list_transactions_api_v1_transactions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get Transaction by ID",
        "description": "Fetches full transaction audit record by transaction ID.",
        "operationId": "get_transaction_api_v1_transactions__transaction_id__get",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/evaluate": {
      "post": {
        "tags": [
          "Autonomous Payment Agent"
        ],
        "summary": "Evaluate User Instruction with Autonomous Agent",
        "description": "Evaluates natural language payment instruction through the autonomous payment agent,\nenforcing pre-execution injection scanning, multi-turn crescendo checks, and post-execution PCI DLP redaction.",
        "operationId": "evaluate_agent_instruction_api_v1_agent_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "get": {
        "tags": [
          "Conversational Sessions"
        ],
        "summary": "List Multi-Turn Agent Sessions",
        "description": "Returns summaries of all active and historical multi-turn conversational agent sessions.",
        "operationId": "list_sessions_api_v1_sessions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max sessions to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max sessions to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummaryResponse"
                  },
                  "title": "Response List Sessions Api V1 Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "tags": [
          "Conversational Sessions"
        ],
        "summary": "Get Session Detail and Turn History",
        "description": "Returns full conversation turn history, risk escalation trajectory, and lock status for a session.",
        "operationId": "get_session_detail_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/security/scan-prompt": {
      "post": {
        "tags": [
          "Security & Threat Detection"
        ],
        "summary": "Dedicated Prompt Threat Scan",
        "description": "Fast-path scan focusing specifically on user prompt injection,\njailbreak attempts, and financial instruction manipulation.",
        "operationId": "scan_prompt_api_v1_security_scan_prompt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/security/scan-response": {
      "post": {
        "tags": [
          "Security & Threat Detection"
        ],
        "summary": "DLP and PAN Response Scanner",
        "description": "Outbound response DLP scanner. Detects, validates with Luhn,\nand masks PANs, CVVs, and credentials before user delivery.",
        "operationId": "scan_response_api_v1_security_scan_response_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/security/events/{event_id}": {
      "get": {
        "tags": [
          "Security & Threat Detection"
        ],
        "summary": "Get Security Event by ID",
        "description": "Retrieves full details of a specific security event.",
        "operationId": "get_security_event_api_v1_security_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "Alerts & Incident Response"
        ],
        "summary": "List Security Alerts",
        "description": "Lists alerts with active count and resolution status filtering.",
        "operationId": "list_alerts_api_v1_alerts_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "is_resolved",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by resolution status",
              "title": "Is Resolved"
            },
            "description": "Filter by resolution status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/resolve": {
      "post": {
        "tags": [
          "Alerts & Incident Response"
        ],
        "summary": "Resolve Alert",
        "description": "Marks an active security alert as investigated and resolved by an analyst.",
        "operationId": "resolve_alert_api_v1_alerts__alert_id__resolve_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertResolveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/simulation/run": {
      "post": {
        "tags": [
          "Traffic & Attack Simulation"
        ],
        "summary": "Run Payment Traffic or Attack Simulation",
        "description": "Executes a reproducible batch of simulated transactions (normal, adversarial, or red-team)\nthrough the live layered security engine and commits audit events.",
        "operationId": "run_simulation_api_v1_simulation_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Run Simulation Api V1 Simulation Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AgentDecisionResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "turn_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Index"
          },
          "source_currency": {
            "type": "string",
            "title": "Source Currency"
          },
          "destination_currency": {
            "type": "string",
            "title": "Destination Currency"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "exchange_rate": {
            "type": "number",
            "title": "Exchange Rate"
          },
          "destination_country": {
            "type": "string",
            "title": "Destination Country"
          },
          "route": {
            "type": "string",
            "title": "Route"
          },
          "decision": {
            "type": "string",
            "enum": [
              "APPROVE",
              "FLAG",
              "REJECT",
              "BLOCKED"
            ],
            "title": "Decision"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "agent_risk_score": {
            "type": "number",
            "title": "Agent Risk Score"
          },
          "raw_response": {
            "type": "string",
            "title": "Raw Response"
          },
          "tool_calls": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Calls"
          },
          "execution_latency_ms": {
            "type": "number",
            "title": "Execution Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "source_currency",
          "destination_currency",
          "amount",
          "exchange_rate",
          "destination_country",
          "route",
          "decision",
          "reason",
          "agent_risk_score",
          "raw_response",
          "execution_latency_ms"
        ],
        "title": "AgentDecisionResponse"
      },
      "AgentEvaluateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Client or session ID",
            "default": "usr_agent_01"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Optional session ID for stateful multi-turn conversational tracking"
          },
          "user_prompt": {
            "type": "string",
            "title": "User Prompt",
            "description": "Natural language instruction for the autonomous payment agent"
          },
          "context_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Metadata"
          }
        },
        "type": "object",
        "required": [
          "user_prompt"
        ],
        "title": "AgentEvaluateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionSummaryResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "total_turns": {
            "type": "integer",
            "title": "Total Turns"
          },
          "total_violations": {
            "type": "integer",
            "title": "Total Violations"
          },
          "session_locked": {
            "type": "boolean",
            "title": "Session Locked"
          },
          "lock_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lock Reason"
          },
          "avg_risk_score": {
            "type": "number",
            "title": "Avg Risk Score"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_id",
          "total_turns",
          "total_violations",
          "session_locked",
          "avg_risk_score",
          "created_at"
        ],
        "title": "SessionSummaryResponse"
      },
...
      "TransactionCreateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Unique client/user identifier"
          },
          "source_currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Source Currency",
            "description": "ISO source currency code"
          },
          "destination_currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Destination Currency",
            "description": "ISO destination currency code"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "Amount to send"
          },
          "destination_country": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Destination Country",
            "description": "ISO 2-letter country code"
          },
          "route": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route",
            "description": "Corridor rail: SEPA, SWIFT, ACH, FEDNOW, etc.",
            "default": "SEPA"
          },
          "user_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Prompt",
            "description": "Optional natural language instruction"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "source_currency",
          "destination_currency",
          "amount",
          "destination_country"
        ],
        "title": "TransactionCreateRequest"
      },
      "TransactionListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "TransactionListResponse"
      },
      "TransactionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "source_currency": {
            "type": "string",
            "title": "Source Currency"
          },
          "destination_currency": {
            "type": "string",
            "title": "Destination Currency"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "exchange_rate": {
            "type": "number",
            "title": "Exchange Rate"
          },
          "destination_amount": {
            "type": "number",
            "title": "Destination Amount"
          },
          "destination_country": {
            "type": "string",
            "title": "Destination Country"
          },
          "route": {
            "type": "string",
            "title": "Route"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "FLAGGED",
              "BLOCKED",
              "REJECTED"
            ],
            "title": "Status"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "execution_latency_ms": {
            "type": "number",
            "title": "Execution Latency Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "source_currency",
          "destination_currency",
          "amount",
          "exchange_rate",
          "destination_amount",
          "destination_country",
          "route",
          "status",
          "risk_score",
          "execution_latency_ms",
          "created_at"
        ],
        "title": "TransactionResponse"
      },
...
    }
  }
}